Update API to schema
