+++ data_tools/schema/_file.py
@@ -1,67 +1,9 @@
 from enum import StrEnum
-from collections.abc import Callable
-from abc import ABC, abstractmethod
 from typing import Union, List, Any
 from functools import reduce
 
 
 type StringCollection = Union[str, List[str]]
-
-
-class UnwrappedError(Exception):
-    pass
-
-
-class Result:
-    """
-    Algebraic datatype that can either represent a successful result of some operation,
-    or a failure.
-
-    A `Result` can be unwrapped to retrieve the result or raise the exception raised wrapped in a `UnwrappedError`.
-    """
-    class ResultType(StrEnum):
-        Ok = "Ok"
-        Err = "Err"
-
-    def __init__(self, result: Any, result_type: ResultType):
-        self._result = result
-        self._result_type = result_type
-
-    @staticmethod
-    def Ok(result: Any):
-        """
-        Wrap a successful result in a `Result`.
-
-        :param result: the successful result to be wrapped
-        :return: a `Result` instance wrapping `result`.
-        """
-        return Result(result, Result.ResultType.Ok)
-
-    @staticmethod
-    def Err(error: Exception):
-        """
-        Wrap an error/failure/exception in a `Result`.
-
-        :param error: the error to be wrapped
-        :return: a `Result` instance wrapping `error`.
-        """
-        return Result(error, Result.ResultType.Err)
-
-    def unwrap(self):
-        """
-        Unwrap this `Result` to reveal a successful result or an error.
-
-        :raises UnwrappedError: if an error is unwrapped
-        :return: the result, if it was successful
-        """
-        if self._result_type == self.ResultType.Ok:
-            return self._result
-
-        elif self._result_type == self.ResultType.Err:
-            raise UnwrappedError from self._result
-
-    def __bool__(self):
-        return True if self._result_type == self.ResultType.Ok else False
 
 
 class FileType(StrEnum):
@@ -131,46 +73,3 @@ class File:
         :return: a List[str] of path elements
         """
         return canonical_path.split("/")
-
-
-class FileLoader:
-    """
-    A callable object that wraps a lambda function to acquire a `File`, as well as the
-    canonical path that will be queried for the `File` in question.
-    """
-    def __init__(self, loader: Callable[[], Result], canonical_path: str) -> None:
-        """
-        Wrap a `loader` along with the `canonical_path` that it will be querying.
-
-        :param loader: a zero-argument lambda that will return either a `File` or raise a `FileNotFoundError`.
-        :param canonical_path: the path that the lambda will be querying.
-        """
-        self._loader = loader
-        self._canonical_path = canonical_path
-
-    def __call__(self) -> Result:
-        """
-        Invoke this `FileLoader`, and obtain a `Result` containing a File` or an `FileNotFoundError` if it cannot be found.
-
-        :raises FileNotFoundError: If the `File` cannot be loaded
-        :return: the `File` that was loaded
-        """
-        return self._loader()
-
-    @property
-    def canonical_path(self) -> str:
-        return self._canonical_path
-
-
-class DataSource(ABC):
-    def __init__(self, *args, **kwargs) -> None:
-        pass
-
-    @abstractmethod
-    def store(self, **kwargs) -> FileLoader:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get(self, canonical_path: str, **kwargs) -> Result:
-        raise NotImplementedError
-
